use typescript instead of babel & flow
