--- .flowconfig
@@ -1,8 +1,0 @@
-[ignore]
-
-[include]
-
-[libs]
-./interfaces/
-
-[options]
