--- interfaces/ramda.js
@@ -1,5 +1,0 @@
-declare module 'ramda' {
-  declare module.exports: {
-    clamp(rangeStart: string, rangeEnd: number, num: number): number;
-  }
-}
