+++ tslint.json
@@ -1,0 +1,6 @@
+{
+  "extends": "tslint-config-standard",
+  "rules": {
+    "trailing-comma": [true, {"multiline": "always", "singleline": "never"}]
+  }
+}
